AMLS-3873 Changed notification package name from v1 to v1m0.
