+++ app/views/notifications/v1m0/message_details.scala.html
