+++ test/controllers/NotificationControllerSpec.scala
@@ -66,7 +66,7 @@ class NotificationControllerSpec extends AmlsSpec with MockitoSugar with ScalaFu
       receivedAt = new DateTime(2017, 12, 1, 1, 3, DateTimeZone.UTC),
       false,
       amlsRegistrationNumber,
-      "v1",
+      "v1m0",
       IDType("132456")
     )
 
@@ -237,7 +237,7 @@ class NotificationControllerSpec extends AmlsSpec with MockitoSugar with ScalaFu
           "dfgdhsjk",
           ContactType.ApplicationAutorejectionForFailureToPay,
           amlsRegistrationNumber,
-          "v1"
+          "v1m0"
         )(request)
 
         status(result) mustBe 200
@@ -259,7 +259,7 @@ class NotificationControllerSpec extends AmlsSpec with MockitoSugar with ScalaFu
         when(mockNotificationService.getMessageDetails(any(), any(), any())(any(), any()))
           .thenReturn(Future.successful(Some(notificationDetails)))
 
-        val result = controller.messageDetails("dfgdhsjk", ContactType.ReminderToPayForVariation, amlsRegistrationNumber, "v1")(request)
+        val result = controller.messageDetails("dfgdhsjk", ContactType.ReminderToPayForVariation, amlsRegistrationNumber, "v1m0")(request)
 
         status(result) mustBe 200
         contentAsString(result) must include(
@@ -284,7 +284,7 @@ class NotificationControllerSpec extends AmlsSpec with MockitoSugar with ScalaFu
         when(mockNotificationService.getMessageDetails(any(), any(), any())(any(), any()))
           .thenReturn(Future.successful(Some(notificationDetails)))
 
-        val result = controller.messageDetails("id", ContactType.MindedToRevoke, amlsRegistrationNumber, "v1")(request)
+        val result = controller.messageDetails("id", ContactType.MindedToRevoke, amlsRegistrationNumber, "v1m0")(request)
 
         status(result) mustBe 200
         contentAsString(result) must include(msgTxt)
@@ -309,7 +309,7 @@ class NotificationControllerSpec extends AmlsSpec with MockitoSugar with ScalaFu
         when(mockNotificationService.getMessageDetails(any(), any(), any())(any(), any()))
           .thenReturn(Future.successful(Some(notificationDetails)))
 
-        val result = controller.messageDetails("id", ContactType.MindedToReject, amlsRegistrationNumber, "v1")(request)
+        val result = controller.messageDetails("id", ContactType.MindedToReject, amlsRegistrationNumber, "v1m0")(request)
 
         status(result) mustBe 200
         contentAsString(result) must include(msgTxt)
@@ -333,7 +333,7 @@ class NotificationControllerSpec extends AmlsSpec with MockitoSugar with ScalaFu
         when(mockNotificationService.getMessageDetails(any(), any(), any())(any(), any()))
           .thenReturn(Future.successful(Some(notificationDetails)))
 
-        val result = controller.messageDetails("id", ContactType.RejectionReasons, amlsRegistrationNumber, "v1")(request)
+        val result = controller.messageDetails("id", ContactType.RejectionReasons, amlsRegistrationNumber, "v1m0")(request)
 
         status(result) mustBe 200
         contentAsString(result) must include(msgTxt)
@@ -358,7 +358,7 @@ class NotificationControllerSpec extends AmlsSpec with MockitoSugar with ScalaFu
         when(mockNotificationService.getMessageDetails(any(), any(), any())(any(), any()))
           .thenReturn(Future.successful(Some(notificationDetails)))
 
-        val result = controller.messageDetails("id", ContactType.NoLongerMindedToReject, amlsRegistrationNumber, "v1")(request)
+        val result = controller.messageDetails("id", ContactType.NoLongerMindedToReject, amlsRegistrationNumber, "v1m0")(request)
 
         status(result) mustBe 200
         contentAsString(result) must not include msgTxt
@@ -382,7 +382,7 @@ class NotificationControllerSpec extends AmlsSpec with MockitoSugar with ScalaFu
         when(mockNotificationService.getMessageDetails(any(), any(), any())(any(), any()))
           .thenReturn(Future.successful(Some(notificationDetails)))
 
-        val result = controller.messageDetails("id", ContactType.RevocationReasons, amlsRegistrationNumber, "v1")(request)
+        val result = controller.messageDetails("id", ContactType.RevocationReasons, amlsRegistrationNumber, "v1m0")(request)
 
         status(result) mustBe 200
         contentAsString(result) must include(msgTxt)
@@ -408,7 +408,7 @@ class NotificationControllerSpec extends AmlsSpec with MockitoSugar with ScalaFu
         when(mockNotificationService.getMessageDetails(any(), any(), any())(any(), any()))
           .thenReturn(Future.successful(Some(notificationDetails)))
 
-        val result = controller.messageDetails("id", ContactType.NoLongerMindedToRevoke, amlsRegistrationNumber, "v1")(request)
+        val result = controller.messageDetails("id", ContactType.NoLongerMindedToRevoke, amlsRegistrationNumber, "v1m0")(request)
 
         status(result) mustBe 200
         contentAsString(result) must not include msgTxt
@@ -432,7 +432,7 @@ class NotificationControllerSpec extends AmlsSpec with MockitoSugar with ScalaFu
         when(mockNotificationService.getMessageDetails(any(), any(), any())(any(), any()))
           .thenReturn(Future.successful(None))
 
-        val result = controller.messageDetails("", ContactType.MindedToReject, amlsRegistrationNumber, "v1")(request)
+        val result = controller.messageDetails("", ContactType.MindedToReject, amlsRegistrationNumber, "v1m0")(request)
 
         status(result) must be(NOT_FOUND)
       }
@@ -446,12 +446,10 @@ class NotificationControllerSpec extends AmlsSpec with MockitoSugar with ScalaFu
           .thenReturn(Future.successful(statusResponse.copy(safeId = None)))
 
         intercept[Exception]{
-          await(controller.messageDetails("", ContactType.MindedToReject, amlsRegistrationNumber, "v1")(request))
+          await(controller.messageDetails("", ContactType.MindedToReject, amlsRegistrationNumber, "v1m0")(request))
         }.getMessage must be("Unable to retrieve SafeID")
-
       }
     }
-
   }
 
 }
