+++ test/models/notifications/NotificationRowSpec.scala
@@ -36,8 +36,7 @@ class NotificationRowSpec extends PlaySpec with AmlsSpec {
     DateTime.now(),
     false,
     "XJML00000200000",
-    "1",
-    IDType("1234567")
+    _id=IDType("1234567")
   )
 
   "NotificationRows " must {
