+++ test/views/notifications/v1m0/no_longer_minded_to_reject_flattenSpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package views.notifications.v1
+package views.notifications.v1m0
 
 import models.notifications.NotificationParams
 import org.scalatest.MustMatchers
