+++ test/views/notifications/v1m0/revocation_reasonsSpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package views.notifications.v1
+package views.notifications.v1m0
 
 import models.notifications.NotificationParams
 import org.scalatest.MustMatchers
@@ -22,23 +22,23 @@ import play.api.i18n.Messages
 import utils.AmlsSpec
 import views.Fixture
 
-class no_longer_minded_to_rejectSpec extends AmlsSpec with MustMatchers {
+class revocation_reasonsSpec extends AmlsSpec with MustMatchers {
 
   trait ViewFixture extends Fixture {
 
     implicit val requestWithToken = addToken(request)
 
-    val notificationParams = NotificationParams(msgContent = "msgContent", reference = Some("reference"))
+    val notificationParams = NotificationParams(businessName = "Fake Name Ltd.", msgContent = "msgContent", amlsRefNo = "amlsRegNo", endDate = "endDate")
 
   }
 
-  "minded_to_reject view" must {
+  "revocation_reasons view" must {
 
     "have correct title" in new ViewFixture {
 
-      def view = views.html.notifications.v1.no_longer_minded_to_reject(notificationParams)
+      def view = views.html.notifications.v1.revocation_reasons(notificationParams)
 
-      doc.title must be("No longer considering refusal" +
+      doc.title must be("Your supervision has been revoked" +
         " - " + "Your registration" +
         " - " + Messages("title.amls") +
         " - " + Messages("title.gov"))
@@ -46,13 +46,14 @@ class no_longer_minded_to_rejectSpec extends AmlsSpec with MustMatchers {
 
     "have correct headings" in new ViewFixture {
 
-      def view = views.html.notifications.v1.no_longer_minded_to_reject(notificationParams)
+      def view = views.html.notifications.v1.revocation_reasons(notificationParams)
 
-      heading.html must be("No longer considering refusal")
+      heading.html must be("Your supervision has been revoked")
       subHeading.html must include("Your registration")
 
     }
 
   }
+
 
 }
