+++ test/views/notifications/v1m0/your_messagesSpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package views.notifications.v1
+package views.notifications.v1m0
 
 import models.notifications.ContactType.RenewalApproval
 import models.notifications.StatusType.{Approved, Rejected}
